+++ lib/controller/generate_layer_controller.dart
@@ -1,21 +1,32 @@
-import 'package:js_cli/core/utils/directory_utils.dart';
-import 'package:js_cli/core/utils/global_variable.dart';
-import 'package:js_cli/core/utils/output_utils.dart';
-import 'package:js_cli/core/utils/reserved_words.dart';
-import 'package:js_cli/models/entities/design_pattern/controller_design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/datasource_design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/datasource_interface_design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/page_design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/repository_design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/repository_interface_design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/usecase_design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/usecase_interface_design_pattern.dart';
+import 'package:ft_cli/core/utils/directory_utils.dart';
+import 'package:ft_cli/core/utils/global_variable.dart';
+import 'package:ft_cli/core/utils/output_utils.dart';
+import 'package:ft_cli/core/utils/reserved_words.dart';
+import 'package:ft_cli/core/utils/triggers_utils.dart';
+import 'package:ft_cli/models/entities/design_pattern/controller_design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/datasource_design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/datasource_interface_design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/page_design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/repository_design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/repository_interface_design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/usecase_design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/usecase_interface_design_pattern.dart';
 import 'package:path/path.dart';
 
 class GenerateLayerController {
   Future<void> call() async {
     _createDirectories();
+
+    await applyTriggersIfNecessary();
+  }
+
+  static Future applyTriggersIfNecessary() {
+    return TriggersUtils.applyIfNecessary(
+      root: 'template',
+      subPath: 'layer',
+      prefixNameReplaceFile: 'complete',
+    );
   }
 
   void _createDirectories() {
@@ -36,6 +47,8 @@ class GenerateLayerController {
 
     await DirectoryUtils.create(path);
 
-    warn('Pasta criada: diretório de ${designPattern.nameDesignPattern()} $path');
+    warn(
+      'Pasta criada: diretório de ${designPattern.nameDesignPattern()} $path',
+    );
   }
 }