+++ lib/controller/init_controller.dart
@@ -1,19 +1,20 @@
-import 'package:js_cli/controller/design_pattern/design_pattern_controller.dart';
-import 'package:js_cli/core/utils/constants.dart';
-import 'package:js_cli/core/utils/output_utils.dart';
-import 'package:js_cli/models/entities/design_pattern/controller_design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/datasource_design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/datasource_interface_design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/page_design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/repository_design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/repository_interface_design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/usecase_design_pattern.dart';
-import 'package:js_cli/models/entities/design_pattern/usecase_interface_design_pattern.dart';
-import 'package:js_cli/models/entities/microfrontend/base_app.dart';
-import 'package:js_cli/models/entities/microfrontend/micro_commons.dart';
-import 'package:js_cli/models/entities/microfrontend/micro_core.dart';
-import 'package:js_cli/models/entities/microfrontend/microfrontend.dart';
+import 'package:ft_cli/controller/design_pattern/design_pattern_controller.dart';
+import 'package:ft_cli/controller/generate_layer_controller.dart';
+import 'package:ft_cli/core/utils/constants.dart';
+import 'package:ft_cli/core/utils/output_utils.dart';
+import 'package:ft_cli/models/entities/design_pattern/controller_design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/datasource_design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/datasource_interface_design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/page_design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/repository_design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/repository_interface_design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/usecase_design_pattern.dart';
+import 'package:ft_cli/models/entities/design_pattern/usecase_interface_design_pattern.dart';
+import 'package:ft_cli/models/entities/microfrontend/base_app.dart';
+import 'package:ft_cli/models/entities/microfrontend/micro_commons.dart';
+import 'package:ft_cli/models/entities/microfrontend/micro_core.dart';
+import 'package:ft_cli/models/entities/microfrontend/microfrontend.dart';
 
 import 'microfrontend/microfrontend_controller.dart';
 
@@ -37,6 +38,8 @@ class InitController {
     _configMicroFrontend(MicroCore());
     _configMicroFrontend(BaseApp());
     _configMicroFrontend(MicroCommons());
+
+    GenerateLayerController.applyTriggersIfNecessary();
   }
 
   Future<void> _configDesignPattern(DesignPattern designPattern) async {