+++ lib/core/files/generate_scripts.dart
@@ -1,8 +1,8 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:js_cli/core/utils/directory_utils.dart';
-import 'package:js_cli/core/utils/reserved_words.dart';
+import 'package:ft_cli/core/utils/directory_utils.dart';
+import 'package:ft_cli/core/utils/reserved_words.dart';
 import 'package:path/path.dart';
 
 class GenerateScripts {
@@ -12,7 +12,7 @@ class GenerateScripts {
     String prefixNameReplaceFile,
   ) {
     final file = '${prefixNameReplaceFile}_scripts_trigger.json';
-    var root = '.js_cli';
+    var root = '.ft_cli';
 
     path = normalize(ReservedWords.removeWordsInFile(
       fileString: '$group/$path',