+++ lib/core/utils/constants.dart
@@ -1,4 +1,4 @@
 class Constants {
-  static const String packageVersion = '0.1.0';
-  static const String projectName = 'js_cli';
+  static const String packageVersion = '0.1.1';
+  static const String projectName = 'ft_cli';
 }