+++ CHANGELOG.md
@@ -7,8 +7,8 @@
 ## 0.0.4 - bug fix
 - reserved word deal on mac.
 
-## 0.0.2+1 - generate template in .js_cli
--   Now the code generation will generate the file and .js_cli where it can be edited as needed.
+## 0.0.2+1 - generate template in .ft_cli
+-   Now the code generation will generate the file and .ft_cli where it can be edited as needed.
 
 ## 0.0.1 - First Publication
 - Publish package